Add `adjoin` combinator for traversals
This is mostly a translation from adjoin from the haskell lens library: https://hackage.haskell.org/package/lens-5.2/docs/Control-Lens-Unsound.html#v:adjoin

TraversableHomogeneousTuple was added as a way to safely traverse the tuple type that `lensProduct` produces.
An alternative way should be to use the array Traversable, but that requires casting in order for typescript to accept that.